refactor: extract shared wallet selector component for onboarding and login flows
