+++ nt-fe/components/page-component-layout.tsx
@@ -92,26 +92,24 @@ export function PageComponentLayout({
                     </div>
                 </div>
 
-                {!hideLogin && (
-                    <div className="flex items-center gap-3">
-                        <LanguageSwitcher />
-                        <Button
-                            variant="ghost"
-                            size="icon"
-                            onClick={toggleTheme}
-                            aria-label={tHeader("toggleTheme")}
-                            className="h-9 w-9 hover:bg-muted text-muted-foreground hover:text-foreground"
-                        >
-                            {theme === "dark" ? (
-                                <Sun className="h-5 w-5" />
-                            ) : (
-                                <Moon className="h-5 w-5" />
-                            )}
-                        </Button>
+                <div className="flex items-center gap-3">
+                    <LanguageSwitcher />
+                    <Button
+                        variant="ghost"
+                        size="icon"
+                        onClick={toggleTheme}
+                        aria-label={tHeader("toggleTheme")}
+                        className="h-9 w-9 hover:bg-muted text-muted-foreground hover:text-foreground"
+                    >
+                        {theme === "dark" ? (
+                            <Sun className="h-5 w-5" />
+                        ) : (
+                            <Moon className="h-5 w-5" />
+                        )}
+                    </Button>
 
-                        <SignIn />
-                    </div>
-                )}
+                    {!hideLogin && <SignIn />}
+                </div>
             </header>
 
             <main className="flex-1 overflow-y-auto bg-page-bg p-4">
