+++ nt-fe/features/onboarding/components/create-treasury-prompt-modal.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import Link from "next/link";
-import { usePathname } from "next/navigation";
 import { useTranslations } from "next-intl";
 import { Button } from "@/components/button";
 import {
@@ -28,8 +27,7 @@ export function CreateTreasuryPromptModal({
     onCreateTreasury,
 }: CreateTreasuryPromptModalProps) {
     const t = useTranslations("onboarding.createPrompt");
-    const pathname = usePathname();
-    const isOnboardingPath = pathname === "/";
+    const isOnboardingPath = source === "onboarding";
     const descriptionSuffix = isOnboardingPath
         ? t("suffixDemo")
         : t("suffixExploring");
@@ -45,11 +43,12 @@ export function CreateTreasuryPromptModal({
         <Dialog
             open={open}
             onOpenChange={(nextOpen) => {
+                if (!nextOpen && isOnboardingPath) return;
                 onOpenChange(nextOpen);
             }}
         >
             <DialogContent>
-                <DialogHeader className="mb-1">
+                <DialogHeader className="mb-1" closeButton={!isOnboardingPath}>
                     <DialogTitle className="text-left">
                         {t("title")}
                     </DialogTitle>
